refactor: modularize main pages, add dynamic document titles, and clean up codebase
This commit includes:
- Modularization of TraineePage, CompanyPage, and HomePage into smaller reusable components
- Added dynamic document titles based on active tab/route for better UX & SEO
- Cleaned up unused imports, redundant code, and standardized formatting across files
