+++ index.html
@@ -21,7 +21,7 @@
         f.parentNode.insertBefore(j, f);
       })(window, document, "script", "dataLayer", "GTM-TS9QQLV8");
     </script>
-    <title>KADA CONNECT</title>
+    <title>KADA Connect</title>
   </head>
   <body>
     <noscript
