+++ src/components/common/filters/FilterInput.tsx
@@ -1,0 +1,28 @@
+import React from "react";
+
+interface FilterInputProps {
+  label: string;
+  value: string;
+  placeholder?: string;
+  onChange: (value: string) => void;
+}
+
+export const FilterInput: React.FC<FilterInputProps> = ({
+  label,
+  value,
+  placeholder,
+  onChange,
+}) => (
+  <div>
+    <label className="mb-2 block text-xs sm:text-sm font-medium text-gray-700">
+      {label}
+    </label>
+    <input
+      type="text"
+      value={value}
+      onChange={(e) => onChange(e.target.value)}
+      placeholder={placeholder}
+      className="w-full rounded-lg border border-gray-300 px-3 py-2 text-sm focus:ring-2 focus:ring-primary-500 focus:border-primary-500 transition"
+    />
+  </div>
+);
