+++ src/utils/trainees/statusBadgeHelper.ts
